yolo/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
websockets==12.0
ultralytics==8.0.196
opencv-python-headless==4.8.1.78
numpy==1.26.0
Pillow==10.1.0
python-multipart==0.0.6
